python-services/requirements.txt
# Python微服务依赖

# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0

# 日志
loguru==0.7.2

# HTTP客户端
httpx==0.26.0

# YAML处理
pyyaml==6.0.1

# 环境变量
python-dotenv==1.0.0
